# config/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: payment-service-provider
host: 'localhost:8000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /payment:
    x-swagger-router-controller: Payment Service Provier
    post:
      description: Create new payment request
      operationId: createPaymentRequest
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/PaymentRequest'
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/PaymentResponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
  /payment/returnResult:
    x-swagger-router-controller: Receiving payment result
    get:
      description: The url which payment provider will redirect to after payment finished
      operationId: receivingPaymentResult
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Redirect to redirect_url from payment creating api
  /payment/checkTrans:
    x-swagger-router-controller: Check payment transaction by request id
    get:
      description: Query payment transaction
      operationId: checkPaymentTransByRequestId
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: request_id
          required: true
          type: string
        - in: query
          name: reference_id
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/CheckTransReponse'
        default:
          description: Error Occured
          schema:
            $ref: '#/definitions/ErrorResponse'
parameters:
  PaymentRequest:
    name: body
    in: body
    description: The parameters for a payment
    schema:
      $ref: '#/definitions/PaymentRequest'
definitions:
  PaymentRequest:
    required:
      - request_id
      - bank_code
      - total_amount
      - redirect_url
    properties:
      request_id:
        type: string
        example: asfass112343
      bank_code:
        type: string
        example: NCB
        description: The Bank Code
      total_amount:
        type: integer
        example: 100000
        description: Total payment amount
      redirect_url:
        type: string
        example: 'http://example.ononpay.com/redirect_url'
        description: The url will be redirected after a payment finished.
  PaymentResponse:
    required:
      - reference_id
      - request_id
      - status
      - redirect_url
    properties:
      reference_id:
        type: string
        example: fdd2dab0-790c-4d8e-86e6-0c34e27c35c5
        description: The id is generated by service
      request_id:
        type: string
        example: asfass112343
      redirect_url:
        type: string
        example: 'http://123pay.com'
      status:
        type: string
        example: SUCCESS|PENDING|FAIL
      message:
        type: string
        example: Message if status is FAIL
  CheckTransReponse:
    required:
      - reference_id
      - request_id
      - status
    properties:
      reference_id:
        type: string
        example: fdd2dab0-790c-4d8e-86e6-0c34e27c35c5
        description: The id is generated by service
      request_id:
        type: string
        example: asfass112343
      status:
        type: string
        example: SUCCESS|PENDING|FAIL
      message:
        type: string
        example: Message if status is FAIL
  ErrorResponse:
    required:
      - status
    properties:
      status:
        type: string
        example: FAIL
      message:
        type: string
        example: Unknow error